ProjetoFinal/Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
stomp.py==8.1.0
boto3==1.34.0
